replace-omxplayer #comment Move the downloader in media.
